Refactor: Migrate helpers/ and prompts.py to llm/ module
- Move helpers/judge.py → llm/judge.py
- Move helpers/cot.py → llm/cot.py
- Move prompts.py → llm/prompts.py
- Create llm/__init__.py with clean exports
- Update all imports in metrics to use fair_forge.llm
- Remove empty helpers/ directory
- Update CLAUDE.md to reflect final structure

This completes the modular architecture refactoring:
- All LLM-related utilities now in llm/
- Cleaner separation of concerns
- Consistent import paths throughout codebase
