+++ fair_forge/llm/__init__.py
@@ -1,0 +1,5 @@
+"""LLM integration utilities for Fair Forge."""
+from .judge import Judge
+from .cot import CoT
+
+__all__ = ['Judge', 'CoT']
